pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[project]
name = "bq-sync"
version = "0.1.0"
description = "Sync BigQuery resources to a local directory structure."
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "google-cloud-bigquery",
    "google-cloud-bigquery-datatransfer",
    "google-cloud-dataform",
]

[project.scripts]
bq-sync = "bq_sync.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
